refactor: extract file storage to separate class
