+++ embrace-android-delivery/src/main/kotlin/io/embrace/android/embracesdk/internal/delivery/storage/FileStorageService.kt
@@ -1,6 +1,7 @@
 package io.embrace.android.embracesdk.internal.delivery.storage
 
 import io.embrace.android.embracesdk.internal.delivery.StoredTelemetryMetadata
+import io.embrace.android.embracesdk.internal.injection.SerializationAction
 import java.io.InputStream
 
 /**
@@ -9,7 +10,22 @@ import java.io.InputStream
 interface FileStorageService {
 
     /**
-     * Loads a stored payload as an [InputStream]
+     * Stores a payload
+     */
+    fun store(metadata: StoredTelemetryMetadata, action: SerializationAction)
+
+    /**
+     * Deletes a payload
+     */
+    fun delete(metadata: StoredTelemetryMetadata, callback: () -> Unit = {})
+
+    /**
+     * Loads a payload as an [InputStream]
      */
     fun loadPayloadAsStream(metadata: StoredTelemetryMetadata): InputStream?
+
+    /**
+     * Return stored payloads as a list sorted in priority order
+     */
+    fun getStoredPayloads(): List<StoredTelemetryMetadata>
 }
